requirements.txt
# Verified Health Content - Dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0
pyyaml==6.0.2
httpx==0.27.0
feedparser==6.0.11
beautifulsoup4==4.12.3
lxml==5.3.0
youtube-transcript-api==0.6.2
ollama==0.3.3
streamlit==1.38.0
python-dotenv==1.0.1
